requirements.txt
streamlit
pandas
matplotlib
wordcloud
transformers
torch
reportlab
Pillow
